+++ src/ui/about.rs
@@ -39,7 +39,7 @@ impl SimpleComponent for AboutDialog {
             set_website: "https://github.com/an-anime-team/honkers-launcher",
             set_issue_url: "https://github.com/an-anime-team/honkers-launcher/issues",
 
-            set_license_type: gtk::License::Gpl30,
+            set_license_type: gtk::License::Gpl30Only,
             set_version: &APP_VERSION,
 
             set_developers: &[
@@ -100,10 +100,33 @@ impl SimpleComponent for AboutDialog {
 
             set_release_notes_version: &APP_VERSION,
             set_release_notes: &[
+                "<p>Added</p>",
+
+                "<ul>",
+                    "<li>Added \"Indonesia\" wine language option</li>",
+                    "<li>Added writing of the game's output to the \"game.log\" file in the launcher's folder</li>",
+                    "<li>Respect root \".version\" file for game version parsing</li>",
+                "</ul>",
+
                 "<p>Fixed</p>",
 
                 "<ul>",
-                    "<li>Updated background URI API key</li>",
+                    "<li>Fixed \"dwebp\" package name for fedora during initial setup</li>",
+                    "<li>Fixed Discord RPC updates</li>",
+                "</ul>",
+
+                "<p>Changed</p>",
+
+                "<ul>",
+                    "<li>Changed background images processing logic</li>",
+                    "<li>Prioritize parsed game version over the API response</li>",
+                "</ul>",
+
+                "<p>Removed</p>",
+
+                "<ul>",
+                    "<li>Removed \"xdelta3\" dependency</li>",
+                    "<li>Removed migrate installation feature</li>",
                 "</ul>"
             ].join("\n"),