+++ src/ui/preferences/general/mod.rs
@@ -17,14 +17,12 @@ pub mod components;
 
 use components::*;
 
-use crate::ui::migrate_installation::MigrateInstallationApp;
 use crate::ui::preferences::main::PreferencesAppMsg;
 
 use crate::i18n::*;
 use crate::*;
 
 pub struct GeneralApp {
-    migrate_installation: Controller<MigrateInstallationApp>,
     components_page: AsyncController<ComponentsPage>,
 
     game_diff: Option<VersionDiff>,
@@ -51,7 +49,6 @@ pub enum GeneralAppMsg {
     UpdateDownloadedWine,
     UpdateDownloadedDxvk,
 
-    OpenMigrateInstallation,
     RepairGame,
 
     OpenMainPage,
@@ -230,13 +227,6 @@ impl SimpleAsyncComponent for GeneralApp {
                     set_orientation: gtk::Orientation::Horizontal,
                     set_spacing: 8,
                     set_margin_top: 16,
-
-                    gtk::Button {
-                        set_label: &tr!("migrate-installation"),
-                        set_tooltip_text: Some(&tr!("migrate-installation-description")),
-
-                        connect_clicked => GeneralAppMsg::OpenMigrateInstallation
-                    },
 
                     gtk::Button {
                         set_label: &tr!("repair-game"),
@@ -502,10 +492,6 @@ impl SimpleAsyncComponent for GeneralApp {
         tracing::info!("Initializing general settings");
 
         let model = Self {
-            migrate_installation: MigrateInstallationApp::builder()
-                .launch(())
-                .detach(),
-
             components_page: ComponentsPage::builder()
                 .launch(())
                 .forward(sender.input_sender(), std::convert::identity),
@@ -551,14 +537,6 @@ impl SimpleAsyncComponent for GeneralApp {
                 self.components_page.sender()
                     .send(ComponentsPageMsg::UpdateDownloadedDxvk)
                     .unwrap();
-            }
-
-            GeneralAppMsg::OpenMigrateInstallation => unsafe {
-                if let Some(window) = crate::ui::main::PREFERENCES_WINDOW.as_ref() {
-                    self.migrate_installation.widget().set_transient_for(Some(window.widget()));
-                }
-
-                self.migrate_installation.widget().present();
             }
 
             GeneralAppMsg::RepairGame => {